move `migrate-bg-gradient` to core
